+++ packages/api/lib/granule-remove-from-cmr.js
@@ -28,7 +28,7 @@ const models = require('../models');
 const _removeGranuleFromCmr = async (granule, collectionId) => {
   log.info(`granules.removeGranuleFromCmrByGranule granule_id: ${granule.granule_id}, colletion_id: ${collectionId}`);
   if (!granule.published || !granule.cmr_link) {
-    log.info(`Granule ${granule.granule_id} in Collection ${collectionId} is not published to CMR, so cannot be removed from CMR`);
+    log.warn(`Granule ${granule.granule_id} in Collection ${collectionId} is not published to CMR, so cannot be removed from CMR`);
     return;
   }
 
@@ -37,7 +37,9 @@ const _removeGranuleFromCmr = async (granule, collectionId) => {
   const metadata = await cmr.getGranuleMetadata(granule.cmr_link);
 
   // Use granule UR to delete from CMR
-  await cmr.deleteGranule(metadata.title, collectionId);
+  if (metadata) {
+    await cmr.deleteGranule(metadata.title, collectionId);
+  }
 };
 
 /**